+++ app/config.py
@@ -1,0 +1,6 @@
+class Config:
+    SECRET_KEY = 'b\xf0\xf1\xd8\x1d\xab3\x82\xb8\xf58Jp\x0c\x80\x93\xd1\xbf/L\xb2\xd8/\x9d0'
+    MYSQL_HOST = 'localhost'
+    MYSQL_USER = 'root'
+    MYSQL_PASSWORD = 'your_password'
+    MYSQL_DB = 'barbershop_db'
