--- app/templates/cadastro.html
@@ -1,137 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-BR">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Formulário - Cadastro</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            background-color: #f4f4f4;
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            height: 100vh;
-            margin: 0;
-        }
-
-        .container {
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            width: 100%;
-            max-width: 600px;
-            padding: 20px;
-        }
-
-        .form-wrapper {
-            background: #fff;
-            border-radius: 8px;
-            box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-            padding: 20px;
-            width: 100%;
-        }
-
-        h1 {
-            text-align: center;
-            color: #333;
-            margin-bottom: 20px;
-        }
-
-        .form-group {
-            margin-bottom: 15px;
-        }
-
-        label {
-            display: block;
-            font-weight: bold;
-            margin-bottom: 5px;
-            color: #333;
-        }
-
-        input[type="text"],
-        input[type="tel"],
-        input[type="date"],
-        select {
-            width: 100%;
-            padding: 10px;
-            border: 1px solid #ddd;
-            border-radius: 4px;
-            box-sizing: border-box;
-        }
-
-        button {
-            width: 100%;
-            padding: 10px;
-            background-color: #007bff;
-            border: none;
-            border-radius: 4px;
-            color: #fff;
-            font-size: 16px;
-            cursor: pointer;
-            transition: background-color 0.3s, transform 0.3s;
-        }
-
-        button:hover {
-            background-color: #0056b3;
-            transform: scale(1.02);
-        }
-
-        #especialidade-outro {
-            display: none;
-        }
-    </style>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.mask/1.14.16/jquery.mask.min.js"></script>
-    <script>
-        $(document).ready(function() {
-            $('#cpf').mask('000.000.000-00');
-        });
-
-        function toggleOtherField() {
-            var select = document.getElementById('especialidade');
-            var otherField = document.getElementById('especialidade-outro');
-            otherField.style.display = select.value === 'outro' ? 'block' : 'none';
-        }
-    </script>
-</head>
-<body>
-    <div class="container">
-        <div class="form-wrapper">
-            <h1>Formulário de Cadastro</h1>
-            <form action="/submit" method="POST">
-                <div class="form-group">
-                    <label for="nome">Nome Completo:</label>
-                    <input type="text" id="nome" name="nome" required>
-                </div>
-                <div class="form-group">
-                    <label for="cpf">CPF:</label>
-                    <input type="text" id="cpf" name="cpf" placeholder="000.000.000-00" required>
-                </div>
-                <div class="form-group">
-                    <label for="telefone">Telefone:</label>
-                    <input type="tel" id="telefone" name="telefone" required>
-                </div>
-                <div class="form-group">
-                    <label for="data">Data de Nascimento:</label>
-                    <input type="date" id="data" name="data" required>
-                </div>
-                <div class="form-group">
-                    <label for="especialidade">Especialidade do Barbeiro:</label>
-                    <select id="especialidade" name="especialidade" onchange="toggleOtherField()" required>
-                        <option value="">Selecione uma especialidade</option>
-                        <option value="corte">Corte de Cabelo</option>
-                        <option value="barba">Barba</option>
-                        <option value="ambos">Corte e Barba</option>
-                        <option value="outro">Outro</option>
-                    </select>
-                    <input type="text" id="especialidade-outro" name="especialidade_outro" placeholder="Especifique outra" maxlength="30">
-                </div>
-                <div class="form-group">
-                    <button type="submit">Cadastrar</button>
-                </div>
-            </form>
-        </div>
-    </div>
-</body>
-</html>
