--- app/templates/login.css
@@ -1,88 +1,0 @@
-/* Reset básico para garantir uma aparência consistente */
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body {
-    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-    background: linear-gradient(135deg, #f3f4f6, #e5e7eb);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-}
-
-.container {
-    width: 100%;
-    max-width: 360px;
-    background: #ffffff;
-    padding: 40px;
-    border-radius: 12px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-}
-
-.form-container h1 {
-    text-align: center;
-    margin-bottom: 30px;
-    font-size: 24px;
-    color: #333;
-}
-
-form {
-    display: flex;
-    flex-direction: column;
-}
-
-form label {
-    margin-bottom: 5px;
-    font-weight: bold;
-    color: #555;
-}
-
-form input {
-    margin-bottom: 20px;
-    padding: 15px;
-    border: 1px solid #ddd;
-    border-radius: 8px;
-    font-size: 16px;
-    transition: border-color 0.3s;
-}
-
-form input:focus {
-    border-color: #4a90e2;
-    outline: none;
-}
-
-form button {
-    padding: 15px;
-    background-color: #4a90e2;
-    border: none;
-    border-radius: 8px;
-    color: white;
-    font-size: 16px;
-    cursor: pointer;
-    transition: background-color 0.3s, transform 0.3s;
-}
-
-form button:hover {
-    background-color: #357abd;
-    transform: scale(1.02);
-}
-
-form p {
-    text-align: center;
-    margin-top: 10px;
-    font-size: 14px;
-}
-
-form p a {
-    color: #4a90e2;
-    text-decoration: none;
-    font-weight: bold;
-}
-
-form p a:hover {
-    text-decoration: underline;
-}
