--- app/templates/login.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-BR">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Login - Barbearia</title>
-    <link rel="stylesheet" href="login.css">
-</head>
-<body>
-    <div class="container">
-        <div class="form-container">
-            <h1>Bem-vindo à Barbearia</h1>
-            <form>
-                <label for="email">E-mail</label>
-                <input type="email" id="email" placeholder="Digite seu e-mail" required>
-                
-                <label for="senha">Senha</label>
-                <input type="password" id="senha" placeholder="Digite sua senha" required>
-                
-                <button type="submit">Entrar</button>
-                
-                <p>Ainda não tem uma conta? <a href="cadastro.html">Criar Conta</a></p>
-
