--- chave.py
@@ -1,3 +1,0 @@
-import os
-secret_key = os.urandom(24)  # Gera uma chave de 24 bytes (recomendada)
-print(secret_key)
